Refactor: Add pages folder to components file structure
