+++ src/app/(main)/page.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import HomeAdCarousel from "@/components/home/HomeAdCarousel";
+import HomeAdCarousel from "@/components/pages/home/HomeAdCarousel";
 import WhiteSpace from "@/components/common/WhiteSpace";
 
 export default function Home() {
-    return (
-        <div>
-            <HomeAdCarousel />
-            <WhiteSpace height={30} backgroundColor="#ffffff" />
-        </div>
-    );
+  return (
+    <div>
+      <HomeAdCarousel />
+      <WhiteSpace height={30} backgroundColor="#ffffff" />
+    </div>
+  );
 }
