+++ src/app/(product)/category/page.tsx
@@ -1,6 +1,6 @@
-import CategoryFooter from "@/components/category/categoryFooter";
-import CategoryList from "@/components/category/categoryList";
-import ThemeList from "@/components/category/themeList";
+import CategoryFooter from "@/components/pages/category/categoryFooter";
+import CategoryList from "@/components/pages/category/categoryList";
+import ThemeList from "@/components/pages/category/themeList";
 import FloatingLeft from "@/components/common/FloatingLeft";
 import React from "react";
